% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_all.R
\name{abs_gsea_all}
\alias{abs_gsea_all}
\title{GSEA analysis}
\usage{
abs_gsea_all(
  listofTopNamed,
  namescomp,
  ranking_metric,
  geneColname,
  organism_annot,
  keyType,
  readable = TRUE,
  GO_categories,
  GO_minSetSize,
  GO_maxSetSize,
  resultsSummFN,
  saveRda = TRUE,
  saveGMT = TRUE,
  label,
  outputDir,
  rdaDir,
  gmtDir,
  GO_pvalcutoff,
  GO_pAdjustMethod,
  saveTables,
  GOPlots,
  GSEAresGO,
  GSEAresReac,
  Reac_pvalcutoff,
  Reac_pAdjustMethod,
  ReacPlots,
  col_entrez,
  Reac_minSetSize,
  Reac_maxSetSize
)
}
\arguments{
\item{organism_annot}{<- Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{GO_minSetSize}{<- Minimum size of the pathways analysed in the analysis}

\item{GO_maxSetSize}{<- Maximum size of the pathways analysed in the analysis}

\item{GO_pvalcutoff}{<- Thresholds of p-value for GO analysis (in this order: BP, CC, MF)}

\item{GO_pAdjustMethod}{<- Adjustment method for each GO category to analyse (CC, BP, MF) in each comparison}

\item{GSEAresGO}{<-Object resulting from the function 'abs_gsea_GOunfilt'}

\item{GSEAresReac}{<-Object resulting from the function 'abs_gsea_ReactomePAunfilt'}

\item{Reac_pvalcutoff}{<- Thresholds of p-value for GO analysis (in this order: BP, CC, MF)}

\item{Reac_pAdjustMethod}{<- Adjustment method for each GO category to analyse (CC, BP, MF) in each comparison}

\item{Reac_minSetSize}{<- Minimum size of the pathways analysed in the analysis}

\item{Reac_maxSetSize}{<- Maximum size of the pathways analysed in the analysis}

\item{annotPackage}{Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}

\item{GO_comparisons}{<- Comparisons to be analysed, indicated by the corresponding position in 'listofcoef' or 'listofTopNames' (eg. 1:10 or c(1,2,8))}

\item{Reac_comparisons}{<- Comparisons to be analysed, indicated by the corresponding position in 'listofcoef' or 'listofTopNames' (eg. 1:10 or c(1,2,8))}
}
\value{
Creates lists of GO and Reactome unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs GO and Reactome analysis for all the comparisons and store unfiltered results (no pvalue threshold) in lists. Creates dataframes of GO and Reactome results filtered by pvalue are obtained and performs a GO significance analysis, giving some plots with the most enriched GO terms or Reactome pathways.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GO}
\keyword{GSEA}
\keyword{ORA}
\keyword{Reactome}
\keyword{plots}
