% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_ReactomePAfiltplot.R
\name{abs_gsea_ReactomePAfiltplot}
\alias{abs_gsea_ReactomePAfiltplot}
\title{GO results Filtering and Plots}
\usage{
abs_gsea_ReactomePAfiltplot(
  GSEAresReac,
  Reac_pvalcutoff,
  Reac_pAdjustMethod,
  saveTables,
  ReacPlots,
  label,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{GSEAresReac}{Object resulting from the function 'abs_gsea_ReactomePAunfilt'}

\item{Reac_pvalcutoff}{Numeric vector of p-value thresholds for Reactome Pathway analysis in each comparison. Defaults to 0.05.}

\item{Reac_pAdjustMethod}{Vector of pvalue adjustment method for Reactome Pathway analysis in each comparison (eg. "BH", "bonferroni" or "none")}
}
\value{
Saves dataframes of Reactome results filtered by pvalue and plots to visualize the results.
}
\description{
Filter GO results according to pvalue threshold and saves a pdf with different plots to visualize the results of enrichment analysis, using clusterProfiler package (dotplot, enrichment map and netplot)
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GSEA}
\keyword{Reactome}
\keyword{plots}
