% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_ReactomePAunfilt.R
\name{abs_gsea_ReactomePAunfilt}
\alias{abs_gsea_ReactomePAunfilt}
\title{Gene Set Enrichment Analysis (GSEA) over Reactome Pathways database}
\usage{
abs_gsea_ReactomePAunfilt(
  listofTopNamed,
  namescomp,
  ranking_metric = "logFC",
  col_entrez = "EntrezID",
  readable = TRUE,
  organism,
  organism_annot,
  Reac_minSetSize = 3,
  Reac_maxSetSize = 500,
  resultsSummFN = "ResultsSummary_ABS-GSEA.txt",
  saveRda = TRUE,
  saveGMT = TRUE,
  label = "",
  outputDir,
  rdaDir,
  gmtDir
)
}
\arguments{
\item{listofTopNamed}{Named list of toptables}

\item{namescomp}{Name of the comparisons to be analysed, corresponding to names in list of toptables}

\item{ranking_metric}{Column from toptable by which features will be ranked}

\item{col_entrez}{Column from toptable containing EntrezID}

\item{readable}{Whether features should be converted to gene symbol. Defaults to TRUE.}

\item{organism}{Organism (eg. 'human', 'rat', 'mouse')}

\item{organism_annot}{Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{Reac_minSetSize}{Minimum size of the pathways analysed in the analysis}

\item{Reac_maxSetSize}{Maximum size of the pathways analysed in the analysis}
}
\value{
Creates lists of Reactome unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs GSEA analysis using clusterProfiler package over Reactome Pathways database for the results from differential expression analysis ranked by the specified metric. The unfiltered enrichment results are returned and saved as xls files.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GSEA}
\keyword{Reactome}
