% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_ora_all.R
\name{abs_ora_all}
\alias{abs_ora_all}
\title{ORA analysis}
\usage{
abs_ora_all(
  listofTopNamed,
  namescomp,
  universe,
  geneColname,
  readable = TRUE,
  selected.pval.go,
  selected.pvalthr.go,
  selected.logFC.go,
  col_logFC.go,
  sign_logFC.go,
  organism_annot,
  keyType,
  GO_categories,
  GO_minSetSize,
  GO_maxSetSize,
  resultsSummFN,
  saveRda,
  saveGMT,
  label,
  outputDir,
  rdaDir,
  gmtDir,
  col_entrez,
  selected.pval.reac,
  selected.pvalthr.reac,
  selected.logFC.reac,
  col_logFC.reac,
  sign_logFC.reac,
  organism,
  Reac_minSetSize,
  Reac_maxSetSize,
  ORAresGO,
  sign,
  GO_pvalcutoff,
  GO_pAdjustMethod,
  saveTables,
  GOPlots,
  ORAresReac,
  Reac_pvalcutoff,
  Reac_pAdjustMethod,
  ReacPlots
)
}
\arguments{
\item{universe}{<- Universe}

\item{geneColname}{<- Name of the column used for the overrepresentation analysis (e.g. "EntrezID")}

\item{selected.pval.go}{<- "P.Value" or "adj.P.Val" (for each set of comparisons)}

\item{selected.pvalthr.go}{<- Pvalue threshold  (for each set of comparisons)}

\item{selected.logFC.go}{<- LogFC threshold}

\item{col_logFC.go}{<- Name of the column where the logFC is located}

\item{sign_logFC.go}{<- Label for deciding what kind of analysis perform: "up", "down" or "abs"}

\item{organism_annot}{<- Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{keyType}{<- ID used in the analysis for gene classification (e.g. "UNIPROT", "ENTREZID")}

\item{GO_categories}{<- List of categories that will be used in the analysis (e.g. "BP", "CC" and/or "MF")}

\item{GO_minSetSize}{<- Minimum set size for the analysis}

\item{GO_maxSetSize}{<- Maximum set size for the analysis}

\item{resultsSummFN}{<- Name of the file where a summary of the results will be saved}

\item{label}{<- Label to use in some file names}

\item{organism}{<- Organism ("human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly")}

\item{ORAresGO}{<- Object resulting from the function 'abs_ORA_GOunfilt'}

\item{sign}{<- vector of length namescomp composed of one number for comparison: +1 for up -1 for down or 0 for abs}

\item{GO_pvalcutoff}{<- Thresholds of p-value for GO analysis (in this order: BP, CC, MF)}

\item{GO_pAdjustMethod}{<- Adjustment method for each GO category to analyse (CC, BP, MF) in each comparison}

\item{GOPlots}{<- Whether to generate plots or not}

\item{ORAresReac}{<- Object resulting from the function 'abs_ORA_ReactomePAunfilt'}

\item{Reac_pvalcutoff}{<- Thresholds of p-value for Reactome overrepresentation analysis (in this order: BP, CC, MF)}

\item{Reac_pAdjustMethod}{<- Adjustment method for each Reactome category to analyse (CC, BP, MF) in each comparison}

\item{ReacPlots}{<- Whether to generate plots or not}

\item{annotPackage}{<- Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}
}
\value{
Creates lists of GO unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs GO and Reactome overrepresentation analysis for all the comparisons and store unfiltered results (no pvalue threshold) in lists. Creates dataframes of GO and Reactome results filtered by pvalue and performs a GO and Reactome significance analysis, giving some plots with the most enriched GO terms and Reactome pathways.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GO}
\keyword{ORA}
\keyword{Reactome}
\keyword{plots}
