% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_ora_GOunfilt.R
\name{abs_ora_GOunfilt}
\alias{abs_ora_GOunfilt}
\title{GO unfilt ORA}
\usage{
abs_ora_GOunfilt(
  listselgenesNamed = NULL,
  listofTopNamed,
  namescomp,
  universe,
  geneColname,
  readable,
  selected.pval.go,
  selected.pvalthr.go,
  selected.logFC.go,
  col_logFC.go,
  sign_logFC.go,
  organism_annot,
  keyType,
  GO_categories,
  GO_minSetSize,
  GO_maxSetSize,
  resultsSummFN,
  saveRda,
  saveGMT,
  label,
  outputDir,
  rdaDir,
  gmtDir
)
}
\arguments{
\item{universe}{<- Universe}

\item{geneColname}{<- Name of the column used for the overrepresentation analysis (e.g. "EntrezID")}

\item{selected.pval.go}{<- "P.Value" or "adj.P.Val" (for each set of comparisons)}

\item{selected.pvalthr.go}{<- Pvalue threshold  (for each set of comparisons)}

\item{selected.logFC.go}{<- LogFC threshold}

\item{col_logFC.go}{<- Name of the column where the logFC is located}

\item{sign_logFC.go}{<- Label for deciding what kind of analysis perform: "up", "down" or "abs"}

\item{organism_annot}{<- Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{keyType}{<- ID used in the analysis for gene classification (e.g. "UNIPROT", "ENTREZID")}

\item{GO_categories}{<- List of categories that will be used in the analysis (e.g. "BP", "CC" and/or "MF")}

\item{GO_minSetSize}{<- Minimum set size for the analysis}

\item{GO_maxSetSize}{<- Maximum set size for the analysis}

\item{resultsSummFN}{<- Name of the file where a summary of the results will be saved}

\item{label}{<- Label to use in some file names}

\item{annotPackage}{<- Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}

\item{organism}{<- Organism ("human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly")}
}
\value{
Creates lists of GO unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs GO overrepresentation analysis for all the comparisons and store unfiltered results (no pvalue threshold) in lists.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GO}
\keyword{ORA}
