% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_ora_ReactomePAfiltplot.R
\name{abs_ora_ReactomePAfiltplot}
\alias{abs_ora_ReactomePAfiltplot}
\title{Reactome filt ORA}
\usage{
abs_ora_ReactomePAfiltplot(
  ORAresReac,
  sign,
  Reac_pvalcutoff,
  Reac_pAdjustMethod,
  saveTables,
  ReacPlots,
  label,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{ORAresReac}{<- Object resulting from the function 'abs_ORA_ReactomePAunfilt'}

\item{sign}{<- vector of length namescomp composed of one number for comparison: +1 for up -1 for down or 0 for abs}

\item{Reac_pvalcutoff}{<- Thresholds of p-value for Reactome overrepresentation analysis (in this order: BP, CC, MF)}

\item{Reac_pAdjustMethod}{<- Adjustment method for each Reactome category to analyse (CC, BP, MF) in each comparison}

\item{ReacPlots}{<- Whether to generate plots or not}

\item{label}{<- Label to use in some file names}

\item{resultsSummFN}{<- Name of the file where a summary of the results will be saved}
}
\value{
Creates dataframes of Reactome results filtered by pvalue and performs a Reactome overrespresentation analysis, giving some plots with the most overrepresented Reactome pathways.
}
\description{
Creates dataframes of Reactome results filtered by pvalue and performs a Reactome overrepresentation analysis, giving some plots with the most enriched GO terms pathways.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{ORA}
\keyword{Reactome}
\keyword{plots}
