% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_ora_ReactomePAunfilt.R
\name{abs_ora_ReactomePAunfilt}
\alias{abs_ora_ReactomePAunfilt}
\title{Reactome unfilt ORA}
\usage{
abs_ora_ReactomePAunfilt(
  listselgenesNamed = NULL,
  listofTopNamed,
  namescomp,
  universe,
  col_entrez,
  readable,
  selected.pval.reac,
  selected.pvalthr.reac,
  selected.logFC.reac,
  col_logFC.reac,
  sign_logFC.reac,
  organism,
  organism_annot,
  keyType,
  Reac_minSetSize,
  Reac_maxSetSize,
  resultsSummFN,
  saveRda,
  saveGMT,
  label,
  outputDir,
  rdaDir,
  gmtDir
)
}
\arguments{
\item{universe}{<- Universe}

\item{col_entrez}{<- Name of the column used for the overrepresentation analysis (e.g. "EntrezID")}

\item{selected.pval.reac}{<- "P.Value" or "adj.P.Val" (for each set of comparisons)}

\item{selected.pvalthr.reac}{<- Pvalue threshold  (for each set of comparisons)}

\item{selected.logFC.reac}{<- LogFC threshold}

\item{organism}{<- Organism ("human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly")}

\item{organism_annot}{<- Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{keyType}{<- ID used in the analysis for gene classification (e.g. "UNIPROT", "ENTREZID")}

\item{Reac_minSetSize}{<- Minimum set size for the analysis}

\item{Reac_maxSetSize}{<- Maximum set size for the analysis}

\item{resultsSummFN}{<- Name of the file where a summary of the results will be saved}

\item{label}{<- Label to use in some file names}

\item{annotPackage}{<- Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}

\item{col_logFC}{<- Name of the column where the logFC is located}

\item{sign_logFC}{<- Label for deciding what kind of analysis perform: "up", "down" or "abs"}
}
\value{
Creates lists of Reactome unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs Reactome overrepresentation analysis for all the comparisons and store unfiltered results (no pvalue threshold) in lists.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{ORA}
\keyword{Reactome}
