% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTargets.R
\name{buildTargets}
\alias{buildTargets}
\title{A function to create a Targets file template}
\usage{
buildTargets(celFilesDir, dataDir, client, ID, descriptors)
}
\arguments{
\item{celFilesDir}{Name of the directory containing the .CEL files.}

\item{dataDir}{Name of the data directory where the targets file will be saved.}

\item{client}{Initials of client}

\item{ID}{Study UEB id}

\item{descriptors}{Variable names to include in dataframe (eg. "FileName","Group", "ShortName", "Colors", "Batch")}
}
\value{
targets A dataframe listing the CEL file names and the covariables
}
\description{
Lists the cel files in a directory and creates a csv2 file containing a dataframe with the file names and phenotypic data to be filled in.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{celfiles}
\keyword{targets}
