% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_all1.R
\name{dea_all1}
\alias{dea_all1}
\title{Create design matrix, fit the linear model and compute contrasts and moderated t-tests}
\usage{
dea_all1(
  data,
  targets,
  sampleNames,
  group,
  covariates,
  fmla = NULL,
  block.cor = NULL,
  contrastsv,
  moderated = TRUE,
  padjust.method = "fdr",
  html_report = FALSE,
  html_ntop = 500,
  html_group = "Group",
  B_thr = c(0),
  Pval_thr = c(0.01, 0.05, 0.1),
  adjPval_thr = c(0.01, 0.05, 0.15, 0.25),
  logFC_thr = 0,
  extended = TRUE,
  B_thr.e = c(0),
  Pval_thr.e = c(0.01, 0.05, 0.1),
  adjPval_thr.e = c(0.01, 0.05, 0.15, 0.25),
  logFC_thr.e = c(0, 0.5, 1, 1.5, 2),
  dovolcanoplot = TRUE,
  volc_logFC = rep(1, length(contrastsv)),
  volc_pval = c(rep("adj.P.Val", length(contrastsv))),
  volc_pval.thr = rep(0.05, length(contrastsv)),
  volc_x0 = rep(-3, length(contrastsv)),
  volc_x1 = rep(+3, length(contrastsv)),
  volc_y0 = rep(0, length(contrastsv)),
  volc_y1 = rep(10, length(contrastsv)),
  n = 6,
  cols = 2,
  label = "",
  summaryFN,
  outputDir
)
}
\arguments{
\item{data}{An ExpressionSet instance or a matrix of expression values, with samples in columns and probes in rows, to which the linear model will be fitted. Column names should be the same and in the same order as the targets sample names used to build the design matrix.}

\item{targets}{Targets dataframe with the phenotypic information of variables and covariates to build the linear model. Can be extracted with pData(eset)}

\item{sampleNames}{Name of variable in targets containing the sample names (the same identificators as the column names of the expression matrix)}

\item{group}{Name of variable in targets containing the main factor of interest (usually: 'Group')}

\item{covariates}{Vector with the name of variable(s) to be used as covariates in the linear model. If not to consider any variable, set NULL.}

\item{fmla}{Formula to be used in the model to construct the design matrix (eg. ' ~ 0 + Group'). Default is NULL. If no formula is provided, it will be automatically built from the main factor and covariates specified in parameters such as: '0 + group + covariates'.}

\item{block.cor}{Name of variable in targets containing the blocking factor accounting for random effects (if it exists) for which an inter-block correlation consensus will be computed and fit to the model. Default is NULL.}

\item{contrastsv}{A character vector with the contrasts to be performed (eg. c("Contrast1=Group2-Group1", "Contrast2=Group3vsGroup1"))}

\item{moderated}{a logical value indicating whether empirical Bayes moderation of the standard errors towards a common value should be performed. Defaults to TRUE. See 'Details' for further information.}

\item{padjust.method}{Method for pvalue adjustment.}

\item{html_report}{true/false}

\item{html_ntop}{number of top genes}

\item{html_group}{group for boxplot}

\item{B_thr}{Numeric vector of thresholds to be used for B statistic}

\item{Pval_thr}{Numeric vector of thresholds to be used for the p-value}

\item{adjPval_thr}{Numeric vector of thresholds to be used for the adjusted p-value}

\item{logFC_thr}{A number indicating the threshold to be used for logFC}

\item{extended}{a logical value indicating whether extended summary tables should be created for each comparison. See \link{dea_summary_ngc_ext} for details.}

\item{B_thr.e}{Numeric vector of thresholds to be used for B statistic in the extended mode}

\item{Pval_thr.e}{Numeric vector of thresholds to be used for the p-value in the extended mode}

\item{adjPval_thr.e}{Numeric vector of thresholds to be used for the adjusted p-value in the extended mode}

\item{logFC_thr.e}{Numeric vector of thresholds to be used for the logFC in the extended mode}

\item{volc_logFC}{Numeric vector of thresholds to be used for logFC (absolute logFC) for each comparison. Genes above this threshold will be colored on the plot.}

\item{volc_pval}{Character vector indicating the name of the variable in the toptable containing the p-values to be used for thresholding (eg. "P.Value" or "adj.P.Val").}

\item{volc_pval.thr}{Numeric vector of thresholds to be used for the p-value for each comparison. Genes above this threshold will be colored on the plot.}

\item{volc_x0}{Numeric vector of edge x inferior limits for each comparison}

\item{volc_x1}{Numeric vector of edge x superior limits for each comparison}

\item{volc_y0}{Numeric vector of edge y inferior limits for each comparison}

\item{volc_y1}{Numeric vector of edge y superior limits for each comparison}

\item{n}{Number of plots per sheet. For multiple plots, n=6 is recommended. Default is 1.}

\item{cols}{Number of columns per sheet. For multiple plots, cols=2 is recommended. Default is 1.}

\item{label}{A string to be included in the file name to be saved}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{outputDir}{Name of the directory where the results summary will be saved}
}
\value{
A list containing the result of the different steps from the differential expression analysis (fit, fit.main, listofcsv, numGenesChanged)
}
\description{
Given a linear model fit to microarray data, compute estimated coefficients and standard errors for a given set of contrasts.
}
\references{

}
\seealso{
\link[maUEB]{dea_lmdesign} \link[maUEB]{dea_lmfit} \link[maUEB]{dea_compare} \link[maUEB]{dea_toptab} \link[maUEB]{dea_summary_ngc1} \link[maUEB]{dea_volcanoplot}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Bayes}
\keyword{contrasts}
\keyword{fit}
\keyword{linear}
\keyword{model}
\keyword{moderated}
\keyword{t-test}
