% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_all2.R
\name{dea_all2}
\alias{dea_all2}
\title{Create design matrix, fit the linear model and compute contrasts and moderated t-tests}
\usage{
dea_all2(
  data,
  targets,
  sampleNames,
  group,
  covariates,
  fmla = NULL,
  block.cor = NULL,
  contrastsv,
  moderated = TRUE,
  summaryFN,
  outputDir,
  ...
)
}
\arguments{
\item{data}{An ExpressionSet instance or a matrix of expression values, with samples in columns and probes in rows, to which the linear model will be fitted. Column names should be the same and in the same order as the targets sample names used to build the design matrix.}

\item{targets}{Targets dataframe with the phenotypic information of variables and covariates to build the linear model. Can be extracted with pData(eset)}

\item{sampleNames}{Name of variable in targets containing the sample names (the same identificators as the column names of the expression matrix)}

\item{group}{Name of variable in targets containing the main factor of interest (usually: 'Group')}

\item{covariates}{Vector with the name of variable(s) to be used as covariates in the linear model. If not to consider any variable, set NULL.}

\item{fmla}{Formula to be used in the model to construct the design matrix (eg. ' ~ 0 + Group'). Default is NULL. If no formula is provided, it will be automatically built from the main factor and covariates specified in parameters such as: '0 + group + covariates'.}

\item{block.cor}{Name of variable in targets containing the blocking factor accounting for random effects (if it exists) for which an inter-block correlation consensus will be computed and fit to the model. Default is NULL.}

\item{contrastsv}{A character vector with the contrasts to be performed (eg. c("Contrast1=Group2-Group1", "Contrast2=Group3vsGroup1"))}

\item{moderated}{a logical value indicating whether empirical Bayes moderation of the standard errors towards a common value should be performed. Defaults to TRUE. See 'Details' for further information.}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{outputDir}{Name of the directory where the results summary will be saved}

\item{...}{Other arguments}
}
\value{
An MArrayLM object containing the result of the fits.
}
\description{
Given a linear model fit to microarray data, compute estimated coefficients and standard errors for a given set of contrasts.
}
\references{

}
\seealso{
\link[maUEB]{dea_lmdesign} \link[maUEB]{dea_lmfit} \link[maUEB]{dea_compare}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Bayes}
\keyword{contrasts}
\keyword{fit}
\keyword{linear}
\keyword{model}
\keyword{moderated}
\keyword{t-test}
