% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_compare.R
\name{dea_compare}
\alias{dea_compare}
\title{Compute contrasts and moderated t-tests}
\usage{
dea_compare(fit, contrastsv, design, moderated = TRUE, summaryFN, outputDir)
}
\arguments{
\item{fit}{An MArrayLM object created by lmFit}

\item{contrastsv}{A character vector with the contrasts to be performed (eg. c("Contrast1=Group2-Group1", "Contrast2=Group3vsGroup1"))}

\item{design}{the design matrix of the microarray experiment, with rows corresponding to arrays and columns to coefficients to be estimated.}

\item{moderated}{a logical value indicating whether empirical Bayes moderation of the standard errors towards a common value should be performed. Defaults to TRUE. See 'Details' for further information.}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{outputDir}{Name of the directory where the results summary will be saved}
}
\value{
An MArrayLM object containing the result of the fits.
}
\description{
Given a linear model fit to microarray data, compute estimated coefficients and standard errors for a given set of contrasts.
}
\references{

}
\seealso{
\link[limma]{contrasts.fit} \link[limma]{eBayes}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Bayes}
\keyword{contrasts}
\keyword{fit}
\keyword{linear}
\keyword{model}
\keyword{moderated}
\keyword{t-test}
