% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_lmdesign.R
\name{dea_lmdesign}
\alias{dea_lmdesign}
\title{Design matrix to be used in the linear model}
\usage{
dea_lmdesign(
  targets,
  sampleNames,
  data,
  group,
  covariates,
  fmla = NULL,
  summaryFN,
  outputDir
)
}
\arguments{
\item{targets}{Targets dataframe with the phenotypic information of variables and covariates to build the linear model. Can be extracted with pData(eset)}

\item{sampleNames}{Name of variable in targets containing the sample names (the same identificators as the column names of the expression matrix)}

\item{data}{An ExpressionSet instance or a matrix of expression values, with samples in columns and probes in rows, to which the linear model will be fitted. Column names should be the same and in the same order as the targets sample names used to build the design matrix.}

\item{group}{Name of variable in targets containing the main factor of interest (usually: 'Group')}

\item{covariates}{Vector with the name of variable(s) to be used as covariates in the linear model. If not to consider any variable, set NULL.}

\item{fmla}{Formula to be used in the model to construct the design matrix (eg. ' ~ 0 + Group'). Default is NULL. If no formula is provided, it will be automatically built from the main factor and covariates specified in parameters such as: '0 + group + covariates'.}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{outputDir}{Name of the directory where the results summary will be saved}
}
\value{
A design matrix
}
\description{
Constructs a design matrix from the main factor and covariates specified in parameters. A formula can be directly provided. The matrix is constructed by expanding factors to a set of dummy variables resulting in a matrix with rows corresponding to samples and columns to factors.
}
\references{

}
\seealso{
\link[stats]{model.matrix}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{design}
\keyword{linear}
\keyword{matrix}
\keyword{model}
