% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_summary_ngc.R
\name{dea_summary_ngc}
\alias{dea_summary_ngc}
\title{Summarise results from differential expression analysis}
\usage{
dea_summary_ngc(
  listofcsv,
  listofcoef,
  B_thr = c(0),
  Pval_thr = c(0.01, 0.05, 0.1),
  adjPval_thr = c(0.01, 0.05, 0.15, 0.25),
  logFC_thr = 0,
  reg = c("Up", "Down"),
  outputDir
)
}
\arguments{
\item{listofcsv}{List of toptables to summarize}

\item{listofcoef}{Name of comparisons, in the same order as listofcsv}

\item{B_thr}{Numeric vector of thresholds to be used for B statistic}

\item{Pval_thr}{Numeric vector of thresholds to be used for the p-value}

\item{adjPval_thr}{Numeric vector of thresholds to be used for the adjusted p-value}

\item{logFC_thr}{A number indicating the threshold to be used for logFC}

\item{reg}{A character vector indicating the direction of logFC change ("Up", "Down", and/or "All"). Default is c("Up", "Down"). If 'All' is indicated, it will take the absolute value of logFC.}

\item{outputDir}{Name of the directory where the results summary will be saved}
}
\value{
A dataframe with summaries for each toptable
}
\description{
Create summary table of number of genes changed in the different contrasts under different thresholds
}
\references{

}
\seealso{
\link{dea_summary_ngc_ext}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{numGenesChanged}
