% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_summary_ngc_ext.R
\name{dea_summary_ngc_ext}
\alias{dea_summary_ngc_ext}
\title{Summarise results from differential expression analysis}
\usage{
dea_summary_ngc_ext(
  listofcsv,
  listofcoef,
  B_thr.e = c(0),
  Pval_thr.e = c(0.01, 0.05, 0.1),
  adjPval_thr.e = c(0.01, 0.05, 0.15, 0.25),
  logFC_thr.e = c(0, 0.5, 1, 1.5, 2),
  outputDir
)
}
\arguments{
\item{listofcsv}{List of toptables to summarize}

\item{listofcoef}{Name of comparisons, in the same order as listofcsv}

\item{B_thr.e}{Numeric vector of thresholds to be used for B statistic in the extended mode}

\item{Pval_thr.e}{Numeric vector of thresholds to be used for the p-value in the extended mode}

\item{adjPval_thr.e}{Numeric vector of thresholds to be used for the adjusted p-value in the extended mode}

\item{logFC_thr.e}{Numeric vector of thresholds to be used for the logFC in the extended mode}

\item{outputDir}{Name of the directory where the results summary will be saved}
}
\value{
A list with summaries for each toptable
}
\description{
Create summary table of number of genes changed in the different contrasts under different thresholds
}
\references{

}
\seealso{
\link{dea_summary_ngc}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{numGenesChanged}
