% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_toptab.R
\name{dea_toptab}
\alias{dea_toptab}
\title{Save toptables}
\usage{
dea_toptab(
  listofcoef,
  fit.main,
  eset,
  padjust.method = "fdr",
  html_report = FALSE,
  html_ntop = 500,
  html_group = "Group",
  outputDir
)
}
\arguments{
\item{listofcoef}{Coefficient names}

\item{fit.main}{MArrayLM object resulting from linear model fit}

\item{eset}{ExpressionSet object containing expression values used to fit the model}

\item{padjust.method}{Method for pvalue adjustment.}

\item{html_report}{true/false}

\item{html_ntop}{number of top genes}

\item{html_group}{group for boxplot}

\item{outputDir}{Name of the directory where the results summary will be saved}

\item{html_padjust_method}{pval to show in htmlreport}
}
\value{
A list with csv2top
}
\description{
Given a linear model fit to microarray data, compute estimated coefficients and standard errors for a given set of contrasts.
}
\references{

}
\seealso{
\link[limma]{topTable} \link[ReportingTools]{HTMLReport}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{topTable}
