% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_volcanoplot.R
\name{dea_volcanoplot}
\alias{dea_volcanoplot}
\title{Volcano plot of results from differential expression analysis}
\usage{
dea_volcanoplot(
  listofcsv,
  listofcoef,
  volc_logFC = rep(1, length(listofcoef)),
  volc_pval = c(rep("adj.P.Val", length(listofcoef))),
  volc_pval.thr = rep(0.05, length(listofcoef)),
  volc_x0 = rep(-3, length(listofcoef)),
  volc_x1 = rep(+3, length(listofcoef)),
  volc_y0 = rep(0, length(listofcoef)),
  volc_y1 = rep(10, length(listofcoef)),
  n = 1,
  cols = 1,
  outputDir,
  label = ""
)
}
\arguments{
\item{listofcsv}{List of toptables to summarize}

\item{listofcoef}{Name of comparisons, in the same order as listofcsv}

\item{volc_logFC}{Numeric vector of thresholds to be used for logFC (absolute logFC) for each comparison. Genes above this threshold will be colored on the plot.}

\item{volc_pval}{Character vector indicating the name of the variable in the toptable containing the p-values to be used for thresholding (eg. "P.Value" or "adj.P.Val").}

\item{volc_pval.thr}{Numeric vector of thresholds to be used for the p-value for each comparison. Genes above this threshold will be colored on the plot.}

\item{volc_x0}{Numeric vector of edge x inferior limits for each comparison}

\item{volc_x1}{Numeric vector of edge x superior limits for each comparison}

\item{volc_y0}{Numeric vector of edge y inferior limits for each comparison}

\item{volc_y1}{Numeric vector of edge y superior limits for each comparison}

\item{n}{Number of plots per sheet. For multiple plots, n=6 is recommended. Default is 1.}

\item{cols}{Number of columns per sheet. For multiple plots, cols=2 is recommended. Default is 1.}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{label}{A string to be included in the file name to be saved}
}
\value{
Creates a volcano plot into the graphical device, saves it as pdf.
}
\description{
Creates a volcano plot of the results from differential expression analysis. Dots are colored according to threshold.
}
\references{

}
\seealso{
\link[ggplot2]{ggplot}
}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{volcano}
