% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering of features in an ExpressionSet}
\usage{
filtering(
  data,
  outputFN,
  outputDir,
  summaryFN,
  feature.exclude = "^AFFX",
  require.entrez = TRUE,
  remove.dupEntrez = TRUE,
  var.filter = FALSE,
  var.func = IQR,
  var.cutoff = 0.5,
  filterByQuantile = TRUE,
  require.GOBP = FALSE,
  require.GOCC = FALSE,
  require.GOMF = FALSE
)
}
\arguments{
\item{data}{An instance of ExpressionSet.}

\item{outputFN}{File name of the datasheet saved containing normalized data (csv or xls file), without extension.}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{feature.exclude}{A character vector of regular expressions.
    Feature identifiers (i.e. value of \code{featureNames(eset)})
    that match one of the specified patterns will be filtered out.
    The default value is intended to filter out Affymetrix quality control
    probe sets.}

\item{require.entrez}{If \code{TRUE}, filter out features
    without an Entrez Gene ID annotation. If using an annotation
    package where an identifier system other than Entrez Gene IDs is
    used as the central ID, then that ID will be required instead.}

\item{remove.dupEntrez}{If \code{TRUE} and there are features
      mapping to the same Entrez Gene ID (or equivalent), then the feature with
      the largest value of \code{var.func} will be retained and the
      other(s) removed.}

\item{var.filter}{A logical indicating whether to perform
      filtering based on \code{var.func}.}

\item{var.func}{The function used as the per-feature filtering
    statistic. This function should return a numeric vector of length
    one when given a numeric vector as input.}

\item{var.cutoff}{A numeric value. If \code{var.filter} is TRUE,
    features whose value of \code{var.func} is less than either:
    the \code{var.cutoff}-quantile of all \code{var.func} values
    (if \code{filterByQuantile} is TRUE), or
    \code{var.cutoff} (if \code{filterByQuantile} is FALSE)
    will be removed.}

\item{filterByQuantile}{A logical indicating whether \code{var.cutoff}
    is to be interprested as a quantile of all \code{var.func} values
    (the default), or as an absolute value.}

\item{require.GOBP}{If \code{TRUE}, filter out features
    whose target genes are not annotated to at least one GO term in
    the BP, CC or MF ontology, respectively.}

\item{require.GOCC}{If \code{TRUE}, filter out features
    whose target genes are not annotated to at least one GO term in
    the BP, CC or MF ontology, respectively.}

\item{require.GOMF}{If \code{TRUE}, filter out features
    whose target genes are not annotated to at least one GO term in
    the BP, CC or MF ontology, respectively.}
}
\value{
Returns an ExpressionSet with filtered features. The filtered expression values are saved into a csv/xls file.
}
\description{
Uses 'nsFilter' function to remove features from an ExpressionSet. Data can be filtered by annotation and/or variance.
}
\note{
For memory reasons, only half of the values are plotted.
}
\references{

}
\seealso{
\link[genefilter]{nsFilter}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{filter}
