% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_all.R
\name{mc_all}
\alias{mc_all}
\title{Multiple comparisons analysis: Venn diagrams, Upset plots and Heatmaps}
\usage{
mc_all(
  listofcsv,
  namescomp,
  label,
  colFeat,
  colPval,
  pval,
  colFC,
  FC,
  pltR = TRUE,
  pltPdf = TRUE,
  pltPng = FALSE,
  venn = TRUE,
  eul = FALSE,
  saveTables = TRUE,
  upsetPlot = FALSE,
  saveTables_extended = TRUE,
  colors = rainbow(length(namescomp)),
  trans = 0.5,
  cex1 = 1,
  rotation = 0,
  position = rep(0, length(namescomp)),
  cex2 = 1,
  include,
  hm_dataset,
  targets,
  featureCol = "Gene.Symbol",
  hm_comparNames,
  hm_compar,
  hm_groupexclude = rep(list(c()), length(hm_comparNames)),
  hm_pval,
  hm_pval.thr,
  hm_logFC = "logFC",
  hm_logFC.thr,
  hm_palette = colorRampPalette(c("blue", "white", "red"))(n = 199),
  hm_clustCol = TRUE,
  hm_clustCol.dist = "euclidian",
  hm_clustCol.method = "complete",
  hm_clustRow.cor = TRUE,
  batcheffect = FALSE,
  batchcolName = NULL,
  batchcolName2 = NULL,
  batchNumcolName = NULL,
  hm_plots_interactive = FALSE,
  Venn_plots = TRUE,
  Heatmap_plots = TRUE,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{listofcsv}{list of the dataframes with the toptables}

\item{namescomp}{list with the names of each comparisons}

\item{label}{name that is going to appear in the file}

\item{colFeat}{column by which common genes are searched}

\item{colPval}{column to do the selection (normally Pvalue)}

\item{pval}{p-value by which the selection is done}

\item{colFC}{column referring to logFC (eg. "logFC")}

\item{FC}{logFC value to select genes}

\item{pltR}{to show or not the image in R}

\item{pltPdf}{to make or not the pdf}

\item{venn}{to perform or not the venn diagrams}

\item{eul}{to draw or not euler plot}

\item{upsetPlot}{to perform or not the upset plots. Normally when there are more than 5 comparisons or when this kind of plot is preferred}

\item{colors}{colors to use for the plots (rainbow by defect)}

\item{trans}{transparency of colors}

\item{cex1}{size of the labels}

\item{rotation}{venn global rotation}

\item{position}{vector with the position in grades of each comparison label}

\item{cex2}{size of the numbers}

\item{include}{vector with direction in which genes are going to be selected (up / down / abs) in each multiple comparison}

\item{hm_dataset}{Dataset with expression values to plot on heatmap}

\item{targets}{Targets file}

\item{featureCol}{Column in targets file corresponding to features (eg. "Gene.Symbol")}

\item{hm_comparNames}{name of set of comparisons (without spaces/dots/_). Usually the same as used for Venn (eg. venn_comparNames)}

\item{hm_compar}{list with set of comparisons to select genes to do the Heatmap, positions (or names) according to 'listofcoef' or 'listofTopNames'. Usually the same as used for Venn (eg. venn_compar)}

\item{hm_groupexclude}{samples to exclude in each set of comparisons for Heatmap (length must be equal to number of comparisons). If no sample is to be excluded put c()}

\item{hm_pval}{"P.Value" or "adj.P.Val" (for each set of comparisons)}

\item{hm_pval.thr}{pvalue threshold (for each set of comparisons)}

\item{hm_logFC}{Name of column corresponding to logFC. Default is "logFC"}

\item{hm_logFC.thr}{logFC threshold (for each set of comparisons)}

\item{hm_palette}{colors used in the heatmap plots}

\item{hm_clustCol}{whether to perform hierarchical clustering of samples in  heatmap (column dendogram) (eg. TRUE/FALSE)}

\item{hm_clustCol.dist}{required if hm_clustCol=TRUE: method for distance calculation in hierarchical clustering of samples (default in heatmap.2 is "euclidian")}

\item{hm_clustCol.method}{required if hm_clustCol=TRUE: method for hierarchical clustering (default in heatmap.2 is "complete", other may be "average")}

\item{hm_clustRow.cor}{whether to cluster gene expression based on distance correlation (otherwise distance matrix is calculated based on euclidian distance (default in heatmap.2 function))}

\item{batcheffect}{whether to correct expression values for batch effect (eg. TRUE/FALSE)}

\item{batchcolName}{column name in 'targets' file containing Batch (eg. 'Batch')}

\item{batchcolName2}{column name in 'targets' file containing categorical cofactor to remove from expression data (eg. 'Batch2')}

\item{batchNumcolName}{Vector of numerical variables contained in 'targets' file to be taken as covariates in removeBatcheffect function (eg. 'Day')}

\item{outputDir}{directory in which the results are going to be saved}
}
\value{
Plot Venn diagramm from 2 to 5 comparisons or UpsetR plot if >5 comparisons. Plot heatmap plots.
}
\description{
Plot Venn diagramm from 2 to 5 comparisons or UpsetR plot if >5 comparisons. It also plots heatmaps between the comparisons determined, with the thresholds values (p-value and logFC) of interest.
The Venn and Upset plots function is based in previous work of Ferran Briansó and Miriam Mota ("https://raw.githubusercontent.com/rgonzalo/GitBackupScripts/master/VennDiagrams/CreateVennEuler.R").
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Heatmap}
\keyword{Upset}
\keyword{Venn}
\keyword{diagrams}
\keyword{plots}
