% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_hm.R
\name{mc_hm}
\alias{mc_hm}
\title{Heatmap plots}
\usage{
mc_hm(
  listofcsv,
  hm_dataset,
  targets,
  featureCol = "Gene.Symbol",
  hm_comparNames,
  hm_compar,
  hm_groupexclude = rep(list(c()), length(hm_comparNames)),
  hm_pval,
  hm_pval.thr,
  hm_logFC = "logFC",
  hm_logFC.thr,
  hm_palette = colorRampPalette(c("blue", "white", "red"))(n = 199),
  hm_clustCol = TRUE,
  hm_clustCol.dist = "euclidian",
  hm_clustCol.method = "complete",
  hm_clustRow.cor = TRUE,
  batcheffect = FALSE,
  batchcolName = NULL,
  batchcolName2 = NULL,
  batchNumcolName = NULL,
  hm_fontsize_row = rep(3, length(hm_comparNames)),
  hm_fontsize_col = rep(7, length(hm_comparNames)),
  hm_fontsize = rep(8, length(hm_comparNames)),
  hm_plots_interactive = FALSE,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{listofcsv}{List of toptables}

\item{hm_dataset}{Dataset with expression values to plot on heatmap}

\item{targets}{Targets file}

\item{featureCol}{Column in targets file corresponding to features (eg. "Gene.Symbol")}

\item{hm_comparNames}{name of set of comparisons (without spaces/dots/_). Usually the same as used for Venn (eg. venn_comparNames)}

\item{hm_compar}{list with set of comparisons to select genes to do the Heatmap, positions (or names) according to 'listofcoef' or 'listofTopNames'. Usually the same as used for Venn (eg. venn_compar)}

\item{hm_groupexclude}{samples to exclude in each set of comparisons for Heatmap (length must be equal to number of comparisons). If no sample is to be excluded put c()}

\item{hm_pval}{"P.Value" or "adj.P.Val" (for each set of comparisons)}

\item{hm_pval.thr}{pvalue threshold (for each set of comparisons)}

\item{hm_logFC}{Name of column corresponding to logFC. Default is "logFC"}

\item{hm_logFC.thr}{logFC threshold (for each set of comparisons)}

\item{hm_palette}{colors used in the heatmap plots}

\item{hm_clustCol}{whether to perform hierarchical clustering of samples in  heatmap (column dendogram) (eg. TRUE/FALSE)}

\item{hm_clustCol.dist}{required if hm_clustCol=TRUE: method for distance calculation in hierarchical clustering of samples (default in heatmap.2 is "euclidian")}

\item{hm_clustCol.method}{required if hm_clustCol=TRUE: method for hierarchical clustering (default in heatmap.2 is "complete", other may be "average")}

\item{hm_clustRow.cor}{whether to cluster gene expression based on distance correlation (otherwise distance matrix is calculated based on euclidian distance (default in heatmap.2 function))}

\item{batcheffect}{whether to correct expression values for batch effect (eg. TRUE/FALSE)}

\item{batchcolName}{column name in 'targets' file containing Batch (eg. 'Batch')}

\item{batchcolName2}{column name in 'targets' file containing categorical cofactor to remove from expression data (eg. 'Batch2')}

\item{batchNumcolName}{Vector of numerical variables contained in 'targets' file to be taken as covariates in removeBatcheffect function (eg. 'Day')}

\item{outputDir}{directory in which the results are going to be saved}
}
\value{
Heatmap plots between the comparisons determined, with the thresholds values (p-value and logFC) of interest. Plots are saved in pdf format in the outputDir.
}
\description{
Plot heatmaps between the comparisons determined, with the thresholds values (p-value and logFC) of interest.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Heatmap}
\keyword{plots}
