% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_all.R
\name{qc_all}
\alias{qc_all}
\title{Quality Control plots and report generation for raw or normalized expression data}
\usage{
qc_all(
  data,
  group,
  group.color,
  samplenames,
  factors,
  factorspvca,
  targetsPVCA = NULL,
  pca_scale,
  colorlist,
  batchRemove,
  batchFactors,
  hc.method = "average",
  hc.numclusters = 2,
  label = "",
  inputDir = NULL,
  outputDir,
  summaryFN,
  doboxplot = TRUE,
  dopca = TRUE,
  dopvca = FALSE,
  dohc = TRUE,
  doarrayQMreport = FALSE
)
}
\arguments{
\item{data}{An ExpressionSet object containing expression data, and the targets as phenotypic data.}

\item{group}{Name of the targets variable specifying the group factor to color samples (eg. "Group")}

\item{group.color}{A vector of group colors to be used to colour the bodies of the box plots}

\item{samplenames}{Name of the targets variable specifying the sample names which will be printed under each boxplot}

\item{factors}{A vector with the names of factor variables contained in targets to color samples in pca plots and/or arrayQM report}

\item{factorspvca}{A vector with the names of factor variables contained in targets to be used in pvca}

\item{targetsPVCA}{If a different targets (phenoData) is to be used for PVCA, the name of the new targets csv file from inputDir should be specified here. Defaults to NULL (not different from pData). See 'Details'.}

\item{pca_scale}{a logical value indicating whether the variables in PCA should be scaled to have unit variance before the analysis takes place.}

\item{colorlist}{Vector of RGB colors to be used for factor variables specified in 'factors' other than Group. If is NULL (default), a default colorlist will be used.}

\item{batchRemove}{a logical value indicating whether batch/covariate effects should be removed from the data prior to PCA}

\item{batchFactors}{Vector with the name of variable(s) from targets dataframe indicating the batches/covariates to be adjusted for}

\item{hc.method}{The agglomeration method as used by 'hclust()' function. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{hc.numclusters}{Number of main clusters to highlight by rect}

\item{label}{A string to be included in the file name to be saved}

\item{inputDir}{Name of the directory from where the targetsPVCA csv file should be loaded, if specified.}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}

\item{doboxplot}{a logical value indicating whether boxplot should be created. Defaults to TRUE.}

\item{dopca}{a logical value indicating whether PCA plot should be created. Defaults to TRUE.}

\item{dopvca}{a logical value indicating whether PVCA plot should be created. Defaults to TRUE.}

\item{dohc}{a logical value indicating whether heatmap of sample distances should be created. Defaults to TRUE.}

\item{arrayQMreport}{a logical value indicating whether arrayQM report should be created. Defaults to TRUE.}
}
\value{
Creates different QC plots into the folder specified.
}
\description{
Performs different Quality Control plots and report generation on the given data matrix.
}
\references{

}
\seealso{
\link{qc_boxplot}, \link{qc_hc}, \link{qc_pca}, \link{qc_pvca}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{Quality}
\keyword{control}
