% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_hc.R
\name{qc_hc}
\alias{qc_hc}
\title{A function to create a Heatmap and a dendogram of sample distances}
\usage{
qc_hc(
  data,
  dist.method = "euclidean",
  hclust.method = "average",
  names,
  cexRow = 0.6,
  cexCol = 0.6,
  rect = TRUE,
  numclusters = 2,
  outputDir,
  label
)
}
\arguments{
\item{data}{A matrix of expression values, with samples in columns and probes in rows.}

\item{dist.method}{The distance measure to be used, as used by 'dist()' function. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust.method}{The agglomeration method as used by 'hclust()' function. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{names}{A vector with sample names corresponding to columns of the data matrix (in the same order)}

\item{cexRow}{Size of labels in rows}

\item{cexCol}{Size of labels in columns}

\item{rect}{Whether a rectangle is to be plot on hierarchical tree to highlight main clusters}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{label}{A string to be included in the file name to be saved}

\item{numClusters}{Number of main clusters to highlight by rect}
}
\value{
Creates a heatmap of sample distances and a dendogram of the hierarchical clustering into the graphical device.
}
\description{
Computes sample distances from a matrix of expression values and creates a heatmap of sample distances and a dendogram of the hierarchical clustering. The plots are saved into a pdf file.
}
\references{

}
\seealso{
\link[stats]{hclust}, \link[stats]{dist}, \link[stats]{heatmap}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{clustering}
\keyword{heatmap}
\keyword{hierarchical}
\keyword{samples}
