% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{qc_pca}
\alias{qc_pca}
\title{A function to create a PCA plot from raw or normalized data}
\usage{
qc_pca(
  data,
  scale,
  factors,
  targets,
  col.group = "Colors",
  colorlist = NULL,
  names,
  outputDir,
  label,
  size = 1.5,
  glineas = 0.25
)
}
\arguments{
\item{data}{A matrix of expression values, with samples in columns and probes in rows.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place.}

\item{factors}{A vector with the names of factor variables contained in targets to color samples in pca plots}

\item{targets}{Targets dataframe with the information of factor variables to plot the pca}

\item{col.group}{Name of variable in targets containing the colors for Group}

\item{colorlist}{Vector of RGB colors to be used for factor variables specified in 'factors' other than Group. If is NULL (default), a default colorlist will be used.}

\item{names}{A vector with sample names corresponding to columns of the data matrix (in the same order)}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{label}{A string to be included in the file name to be saved}

\item{size}{Size of labels}

\item{glineas}{Width of lines rising from labels}
}
\value{
Creates a PCA plot into the graphical device and returns a list with the loads of PCA.
}
\description{
Performs a principal components analysis on the given data matrix and creates a plot of the data projected onto the 2-3 first principal components. The plot is saved into a pdf file.
}
\references{

}
\seealso{
\link[stats]{prcomp}, \link[ggplot2]{ggplot}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{PCA}
