% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pca1.R
\name{qc_pca1}
\alias{qc_pca1}
\title{Principal Component Analysis (PCA) from raw or normalized expression data}
\usage{
qc_pca1(
  data,
  scale,
  dim = 2,
  pca2D_factors = c("Group"),
  pca3D_factors = c("Group"),
  targets,
  col.group = "Colors",
  colorlist = NULL,
  names,
  batchRemove = FALSE,
  batchFactors = NULL,
  outputDir,
  label,
  size = 1.5,
  glineas = 0.25
)
}
\arguments{
\item{data}{A matrix of expression values, with samples in columns and probes in rows.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place.}

\item{dim}{A vector with the number of dimensions (2 and/or 3) to project the PCA. Defaults to 2.}

\item{pca2D_factors}{A vector with the names of factor variables contained in targets to color samples in pca 2D plots. Default is "Group".}

\item{pca3D_factors}{A vector with the names of factor variables contained in targets to color samples in pca 3D plots. Default is "Group".}

\item{targets}{Targets dataframe with the information of factor variables to plot the pca}

\item{col.group}{Name of variable in targets containing the colors for Group}

\item{colorlist}{Vector of RGB colors to be used for factor variables specified in 'factors' other than Group. If is NULL (default), a default colorlist will be used.}

\item{names}{A vector with sample names corresponding to columns of the data matrix (in the same order)}

\item{batchRemove}{a logical value indicating whether batch/covariate effects should be removed from the data prior to PCA}

\item{batchFactors}{Vector with the name of variable(s) from targets dataframe indicating the batches/covariates to be adjusted for}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{label}{A label to be included in the file name to be saved}

\item{size}{Size of labels}

\item{glineas}{Width of lines rising from labels}
}
\value{
Creates a PCA plot into the graphical device, saves it as pdf and returns a list with the loads of PCA.
}
\description{
Performs a principal components analysis on the given data matrix and creates a plot of the data projected onto the 2 or 3 first principal components. The plot is saved into a pdf file (2D) or as html (3D).
}
\references{

}
\seealso{
\link[stats]{prcomp}, \link[ggplot2]{ggplot}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{PCA}
