% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pvca.R
\name{qc_pvca}
\alias{qc_pvca}
\title{A function to assess the source of batch effects in a microarray gene expression experiment}
\usage{
qc_pvca(
  data,
  factors,
  targetsPVCA = NULL,
  pct_threshold = 0.6,
  label = NULL,
  inputDir = NULL,
  outputDir,
  summaryFN
)
}
\arguments{
\item{data}{An instance of ExpressionSet.}

\item{factors}{A vector with the names of factor variables to be assessed}

\item{targetsPVCA}{If a different targets is to be used for PVCA, the name of the new targets csv file should be specified here. See 'Details'.}

\item{pct_threshold}{The percentile value of the minimum amount of the variabilities that the selected principal components need to explain}

\item{label}{A string to be included in the file name to be saved}

\item{inputDir}{Name of the directory from where the targetsPVCA csv file should be loaded, if specified.}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{summaryFN}{Name of the file where the reporting summary of execution results will be saved}
}
\value{
Creates a barplot into the graphical device and returns a list with the loads of PCA.
}
\description{
Performs a Principal Variance Component Analysis (PVCA) using 'pvcaBatchAssess' function from 'pvca' package. A barplot is created with the percentage of total variance associated to each factor assessed. The effects associated to each factor are calculated by fitting all sources as random effects including two-way interaction terms in the Mixed Model to select principal components. See 'Details' for extended information.
}
\details{
The approach leverages the strengths of two very popular data analysis methods: first, principal component analysis (PCA) is used to efficiently reduce data dimension while maintaining the majority of the variability in the data, and variance components analysis (VCA) fits a mixed linear model using factors of interest as random effects to estimate and partition the total variability.
}
\note{
At least two factors should be provided. In addition, for each factor, num of levels must be < num of samples/level, if not it gives an error. To overcome this issue you can create a simplified version of targets file with less levels just to create the pvca.
}
\references{

}
\seealso{
\link[pvca]{pvcaBatchAssess}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{PVCA}
