% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_celfiles.R
\name{read_celfiles}
\alias{read_celfiles}
\title{Read cel files and create an ExpressionSet of raw data}
\usage{
read_celfiles(inputDir, targets)
}
\arguments{
\item{inputDir}{Name of the directory containing the .CEL files to read.}

\item{targets}{Targets object containing the filename of the celfiles to read and phenotypic data}
}
\value{
Returns an ExpressionSet containing raw data
}
\description{
Reads the cel files specified in FileName column of targets dataframe and creates an ExpressionSet
}
\references{

}
\seealso{

}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{ExpressionSet}
\keyword{cel}
\keyword{read}
