% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_targets.R
\name{read_targets}
\alias{read_targets}
\title{A function to read targets file and convert categorical variables to factors}
\usage{
read_targets(inputDir, targetsFN, targets.fact)
}
\arguments{
\item{inputDir}{Name of the directory containing the targets file to read.}

\item{targetsFN}{Name of the targets file}

\item{targets.fact}{Covariables in targets file to be treated as factors}
}
\value{
targets An R object containing the targets file with categorical variables set as factors
}
\description{
Reads a targets file in csv2 format and converts categorical variables to factors
}
\references{

}
\seealso{
read.table
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{csv}
\keyword{read}
