% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removebatch.R
\name{removebatch}
\alias{removebatch}
\title{Remove batch effects from expression data}
\usage{
removebatch(data, targets, batchFactors = NULL)
}
\arguments{
\item{data}{A matrix of expression values, with samples in columns and probes in rows.}

\item{targets}{Targets dataframe with the information of batch variables}

\item{batchFactors}{Vector with the name of variable(s) from targets dataframe indicating the batches/covariates to be adjusted for}
}
\value{
Returns an expression matrix with batch/covariate effects removed.
}
\description{
Removes batch/covariate effects from expression data using linear models (removeBatchEffect function from limma package)
}
\note{
This function is not intended to be used prior to linear modelling. For linear modelling, it is better to include the batch factors in the linear model.
}
\references{

}
\seealso{
\link[limma]{removeBatchEffect}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{batch}
