% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_annotations.R
\name{save_annotations}
\alias{save_annotations}
\title{Save probeset annotations}
\usage{
save_annotations(
  data,
  annotPkg,
  outputFN = "Annotations.AllGenes",
  saveHTML = TRUE,
  title = "Annotations for all genes",
  outputDir
)
}
\arguments{
\item{data}{character vector of probe ids}

\item{annotPkg}{Annotation package to be used for annotations}

\item{outputFN}{File name of the datasheet saved containing annotation data, without extension.}

\item{saveHTML}{Whether annotation datasheet should be saved also as html (TRUE/FALSE). Defaults to TRUE.}

\item{title}{Title included in the html file.}

\item{outputDir}{Name of the directory where the files will be saved}
}
\value{
An aafTable object
}
\description{
Uses aafTableAnn function from annaffy package to construct an aafTable object given a set of probe ids. Saves annotation for all summarized probes as csv and html files
}
\note{
In Exon studies, where probes are summarized at the probeset level but not transcript level, the probe annotations cannot be recovered using this function.
}
\references{

}
\seealso{
\link[annaffy]{aafTableAnn}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{annotation}
