% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_GOfiltplot.R
\name{abs_gsea_GOfiltplot}
\alias{abs_gsea_GOfiltplot}
\title{GO filt}
\usage{
abs_gsea_GOfiltplot(
  GSEAresGO,
  GO_pvalcutoff,
  GO_pAdjustMethod,
  saveTables,
  GOPlots,
  label,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{GSEAresGO}{Object resulting from the function 'abs_gsea_GOunfilt'}

\item{GO_pvalcutoff}{Thresholds of p-value for GO analysis (in this order: BP, CC, MF)}

\item{GO_pAdjustMethod}{Adjustment method for each GO category to analyse (CC, BP, MF) in each comparison}
}
\value{
Creates dataframes of GO results filtered by pvalue and performs a GO significance analysis, giving some plots with the most enriched GO terms pathways.
}
\description{
Performs Creates dataframes of GO results filtered by pvalue are obtained and performs a GO significance analysis, giving some plots with the most enriched GO terms pathways.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GO}
\keyword{GSEA}
\keyword{plots}
