% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_GOunfilt.R
\name{abs_gsea_GOunfilt}
\alias{abs_gsea_GOunfilt}
\title{GO unfilt}
\usage{
abs_gsea_GOunfilt(
  listofTopNamed,
  namescomp,
  ranking_metric = "logFC",
  geneColname = "EntrezID",
  annotPackage,
  organism_annot,
  keyType = "ENTREZID",
  readable = TRUE,
  GO_categories = c("BP", "CC", "MF"),
  GO_minSetSize = 3,
  GO_maxSetSize = 500,
  resultsSummFN = "ResultsSummary_ABS-GSEA.txt",
  saveRda = TRUE,
  saveGMT = TRUE,
  label = "",
  outputDir,
  rdaDir,
  gmtDir
)
}
\arguments{
\item{listofTopNamed}{Named list of toptables}

\item{namescomp}{Name of the comparisons to be analysed, corresponding to names in list of toptables}

\item{ranking_metric}{Column from toptable by which features will be ranked}

\item{geneColname}{Column from toptable containing feature ID}

\item{annotPackage}{Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}

\item{organism_annot}{Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}

\item{GO_minSetSize}{Minimum size of the pathways analysed in the analysis}

\item{GO_maxSetSize}{Maximum size of the pathways analysed in the analysis}
}
\value{
Creates lists of GO unfiltered results and a table with number of terms found at different pvalue thresholds for the different categories analyzed.
}
\description{
Performs GO analysis for all the comparisons and store unfiltered results (no pvalue threshold) in lists.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GO}
\keyword{GSEA}
