% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_gsea_ReactomePAfiltplot.R
\name{abs_gsea_ReactomePAfiltplot}
\alias{abs_gsea_ReactomePAfiltplot}
\title{Reactome filt}
\usage{
abs_gsea_ReactomePAfiltplot(
  GSEAresReac,
  Reac_pvalcutoff,
  Reac_pAdjustMethod,
  saveTables,
  ReacPlots,
  label,
  resultsSummFN,
  outputDir
)
}
\arguments{
\item{GSEAresReac}{<-Object resulting from the function 'abs_gsea_ReactomePAunfilt'}

\item{Reac_pvalcutoff}{<- Thresholds of p-value for GO analysis (in this order: BP, CC, MF)}

\item{Reac_pAdjustMethod}{<- Adjustment method for each GO category to analyse (CC, BP, MF) in each comparison}

\item{annotPackage}{Annotation package (Clariom D/S, specific of species; e.g. "clariomdhumantranscriptcluster.db", "clariomshumantranscriptcluster.db", "mta10transcriptcluster.db", "clariomsmousetranscriptcluster.db")}

\item{organism_annot}{<- Organism annotation package (e.g. "org.Hs.eg.db", "org.Mm.eg.db")}
}
\value{
Creates dataframes of Reactome results filtered by pvalue and performs a Reactome significance analysis, giving some plots with the most enriched Reactome pathways.
}
\description{
Performs Creates dataframes of Reactome results filtered by pvalue are obtained and performs a Reactome significance analysis, giving some plots with the most enriched Reactome pathways.
}
\references{

}
\author{
Mireia Ferrer \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{GSEA}
\keyword{Reactome}
\keyword{plots}
