% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{A function to normalize microarray data}
\usage{
normalization(
  data,
  norm.method = "rma",
  exonStudy = FALSE,
  norm.level = "core",
  bkgremove = TRUE,
  annotPkg = annotPackage,
  outputFN = "Normalized.All",
  outputDir
)
}
\arguments{
\item{data}{An ExpressionSet object with the raw data to be normalized.}

\item{norm.method}{Method used for normalization (currently: oligo::rma)}

\item{exonStudy}{If the analysis is to be performed at exon level (TRUE/FALSE). Defaults to FALSE.}

\item{norm.level}{Level of summarization (only for Exon/Gene arrays). Defaults to 'core' for summarization at transcript level. To summarize at probe level use target='probeset'.}

\item{bkgremove}{Whether to perform RMA background correction (TRUE/FALSE). Defaults to TRUE.}

\item{annotPkg}{Annotation package to be used to annotate the FeatureSet generated}

\item{outputFN}{File name of the datasheet saved containing normalized data (csv or xls file), without extension.}

\item{outputDir}{Name of the directory where the datasheet of normalized data will be saved}
}
\value{
eset_norm An R object containing the normalized data (a FeatureSet)
}
\description{
Normalizes data using one of the available methods (see 'Details') and returns a FeatureSet of normalized data. The normalized expression values are saved into a csv/xls file.
}
\details{
Currently using rma method from oligo package.
}
\references{

}
\seealso{
\link[oligo]{rma}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{normalization}
\keyword{rma}
