% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_boxplot.R
\name{qc_boxplot}
\alias{qc_boxplot}
\title{A function to create a boxplot from raw or normalized data}
\usage{
qc_boxplot(
  data,
  group,
  col,
  names,
  outputDir,
  label,
  cex.axis = 0.5,
  cex.lab = 0.7,
  las = 2,
  plotlegend = TRUE,
  legend.posx = 10,
  legend.posy = 13,
  legend.cex = 0.6,
  ...
)
}
\arguments{
\item{data}{A FeatureSet-like object or ExpressionSet object.}

\item{group}{A vector or factor specifying the grouping variable to color samples}

\item{col}{A vector of colors to be used to colour the bodies of the box plots (for each group)}

\item{names}{A vector of sample names which will be printed under each boxplot}

\item{outputDir}{Name of the directory where the plot will be saved}

\item{label}{A string to be included in the file name to be saved}

\item{cex.axis}{Numeric expansion factor for axis tick labels}

\item{cex.lab}{Numeric expansion factor for axis labels}

\item{las}{Direction of axis tick labels}

\item{plotlegend}{Whether to insert legend into plot (TRUE/FALSE)}

\item{legend.posx}{X coordinate of legend position}

\item{legend.posy}{Y coordinate of legend position}

\item{legend.cex}{Numeric expansion factor for legend text}

\item{...}{other arguments to be passed to boxplot function from oligo package, such as graphical parameters}
}
\description{
Uses the boxplot function from package oligo to create a boxplot of raw or normalized data and saves it into a pdf file. For raw data, expression values are transformed to log2 scale. For normalized data, no transformation is applied since they are already in log2 scale.
}
\note{
The boxplot methods for FeatureSet and Expression use a sample (via sample) of the probes/probesets to produce the plot. Therefore, the user interested in reproducibility is advised to use set.seed.
}
\examples{
createBoxplot(data=rawData, group=targets$Group, col=targets$Colors, names=targets$ShortName, outputDir, label="RawData")
}
\references{

}
\seealso{
\link[oligo]{boxplot}
}
\author{
Mireia Ferrer Almirall \email{mireia.ferrer.vhir@gmail.com}
}
\keyword{ExpressionSet}
\keyword{FEatureSet}
\keyword{boxplot}
